% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{visit_subjects_dates}
\alias{visit_subjects_dates}
\title{Функция показада начала и конца диапазона с пациентами, хуй знает зачем, я учюсь кароч}
\usage{
visit_subjects_dates(
  df,
  subjects = "system_patient_id",
  dates = "visit_time",
  format_ = "\%Y.\%m.\%d"
)
}
\description{
Функция показада начала и конца диапазона с пациентами, хуй знает зачем, я учюсь кароч
}
